Use clean urls in View::get_url() function
Also adds two flags.  One to force get_url to use old-style urls with
ids, because sometimes an id is required by existing javascript which
parses an href attribute.  The other flag just avoids querying user
and group data (for when the caller decides a clean url is not
critical).

Change-Id: I8758cf2fc566f4c523c297d015c262007f365243
Signed-off-by: Richard Mansfield <richard.mansfield@catalyst.net.nz>