Use clean urls for views listed on my friends page
In order to get the url information for the list of views returned by
get_views(), that function is modified to gather the views in a single
flat list rather than in a structure indexed by user id.  Just before
the view list is returned, the url information is fetched and the user
array is built.

Change-Id: I12dad07c90f8c51614e72c7402c63dc38fccceeb
Signed-off-by: Richard Mansfield <richard.mansfield@catalyst.net.nz>